Use working dir, handle spaces in file names
Need to source the function scripts from the directory where the main
script is called.

Update README to show that when we call the script, we need the entire
filename to be one parameter, so wrap in double quotes.

Add a safety check if there aren't exactly two arguments to the script:
1) output dir
2) file path as a single param even if there's spaces

Switch to JSON output from exiftool: CSV output all appears on one line
and filenames with spaces make things difficult.

Use --raw-output with jq so we don't get values wrapped in literal
double-quotes

Fix space-replacement in the model: remove the trailing slash