Add trim function, trim spaces off model
Some EXIF "model" fields have extra spaces on the end.

${var%%  *}

This finds the first two consecutive spaces followed by anything and
deletes all of it.

Examples:
"a  b cd" -> "a"
"a  "     -> "a"
"  a"     -> ""
" a "     -> " a "

${var% }

This finds the the last space followed by nothing, i.e. ' $' in regular
expressions, and then deletes that space.
This handles the case where there isn't two spaces anywhere.

Examples:
"a "       -> "a"
"a  "      -> "a "
" a"       -> " a"
" a "      -> " a"