Cleaning up comments
I also removed the message for when the file started with a period. I
think this was for Picasa original files; they were giving me grief but
I think putting the file size in the file name fixed that ... I don't
remember.

I thought about abstracting out some of the code for
{first,last}_n_words but they're already pretty simple and I don't want
to get clever.